moving out filesystem to utils
