+++ packages/kit/src/core/make_package/index.js
@@ -3,7 +3,7 @@ import globrex from 'globrex';
 import { createRequire } from 'module';
 import * as path from 'path';
 import { preprocess } from 'svelte/compiler';
-import { mkdirp, rimraf } from '../filesystem/index.js';
+import { mkdirp, rimraf } from '../../utils/filesystem.js';
 
 const essential_files = ['README', 'LICENSE', 'CHANGELOG', '.gitignore', '.npmignore'];
